+++ src/main/java/net/snowflake/client/core/SFArrowResultSet.java
@@ -105,7 +105,7 @@ public class SFArrowResultSet extends SFBaseResultSet implements DataConversionC
    */
   SFArrowResultSet(
       SnowflakeResultSetSerializableV1 resultSetSerializable,
-      SFStatement statement,
+      SFStatementImpl statement,
       boolean sortResult)
       throws SQLException {
     this(resultSetSerializable, statement.getSession().getTelemetryClient(), sortResult);
@@ -122,7 +122,7 @@ public class SFArrowResultSet extends SFBaseResultSet implements DataConversionC
     useSessionTimezone = resultSetSerializable.getUseSessionTimezone();
 
     // update the driver/session with common parameters from GS
-    SessionUtil.updateSfDriverParamValues(this.parameters, statement.getSession());
+    SessionUtil.updateSfDriverParamValues(this.parameters, (SFSessionImpl) statement.getSession());
 
     // if server gives a send time, log time it took to arrive
     if (resultSetSerializable.getSendResultTime() != 0) {
