+++ src/main/java/net/snowflake/client/core/SFSession.java
@@ -5,6 +5,7 @@
 package net.snowflake.client.core;
 
 import net.snowflake.client.jdbc.SnowflakeSQLException;
+import net.snowflake.client.jdbc.SnowflakeType;
 import net.snowflake.client.jdbc.telemetry.Telemetry;
 
 import java.sql.DriverPropertyInfo;
@@ -130,4 +131,8 @@ public interface SFSession {
   void setInjectedDelay(int delay);
 
   boolean isSfSQLMode();
+
+  boolean isResultColumnCaseInsensitive();
+
+  SnowflakeType getTimestampMappedType();
 }
