+++ src/main/java/net/snowflake/client/jdbc/SnowflakeFileTransferConfig.java
@@ -8,6 +8,7 @@ import java.io.InputStream;
 import java.util.Properties;
 import net.snowflake.client.core.OCSPMode;
 import net.snowflake.client.core.SFSession;
+import net.snowflake.client.core.SFSessionImpl;
 
 /**
  * This class manages the parameters to call SnowflakeFileTransferAgent.uploadWithoutConnection()
@@ -21,7 +22,7 @@ public class SnowflakeFileTransferConfig {
   private Properties proxyProperties;
   private String prefix;
   private String destFileName;
-  private SFSession session; // Optional, added for S3 and Azure
+  private SFSessionImpl session; // Optional, added for S3 and Azure
   private String command; // Optional, added for S3 and Azure
 
   public SnowflakeFileTransferConfig(Builder builder) {
@@ -69,7 +70,7 @@ public class SnowflakeFileTransferConfig {
     return destFileName;
   }
 
-  public SFSession getSession() {
+  public SFSessionImpl getSession() {
     return session;
   }
 
@@ -87,7 +88,7 @@ public class SnowflakeFileTransferConfig {
     private Properties proxyProperties = null;
     private String prefix = null;
     private String destFileName = null;
-    private SFSession session = null;
+    private SFSessionImpl session = null;
     private String command = null;
 
     public static Builder newInstance() {
@@ -154,7 +155,7 @@ public class SnowflakeFileTransferConfig {
       return this;
     }
 
-    public Builder setSFSession(SFSession session) {
+    public Builder setSFSession(SFSessionImpl session) {
       this.session = session;
       return this;
     }
