+++ src/main/java/net/snowflake/client/jdbc/SnowflakeResultSetSerializableV1.java
@@ -445,7 +445,7 @@ public class SnowflakeResultSetSerializableV1
    * @throws SnowflakeSQLException if failed to parse the result JSON node
    */
   public static SnowflakeResultSetSerializableV1 create(
-      JsonNode rootNode, SFSession sfSession, SFStatement sfStatement)
+      JsonNode rootNode, SFSessionImpl sfSession, SFStatement sfStatement)
       throws SnowflakeSQLException {
     SnowflakeResultSetSerializableV1 resultSetSerializable = new SnowflakeResultSetSerializableV1();
     logger.debug("Entering create()");
