+++ src/main/java/net/snowflake/client/jdbc/cloud/storage/SnowflakeGCSClient.java
@@ -21,10 +21,8 @@ import java.net.URISyntaxException;
 import java.security.InvalidKeyException;
 import java.util.*;
 import java.util.Map.Entry;
-import net.snowflake.client.core.HttpUtil;
-import net.snowflake.client.core.OCSPMode;
-import net.snowflake.client.core.ObjectMapperFactory;
-import net.snowflake.client.core.SFSession;
+
+import net.snowflake.client.core.*;
 import net.snowflake.client.jdbc.*;
 import net.snowflake.client.log.ArgSupplier;
 import net.snowflake.client.log.SFLogger;
@@ -57,7 +55,7 @@ public class SnowflakeGCSClient implements SnowflakeStorageClient {
   private StageInfo stageInfo;
   private RemoteStoreFileEncryptionMaterial encMat;
   private Storage gcsClient = null;
-  private SFSession session = null;
+  private SFSessionImpl session = null;
 
   private static final SFLogger logger = SFLoggerFactory.getLogger(SnowflakeGCSClient.class);
 
@@ -70,7 +68,7 @@ public class SnowflakeGCSClient implements SnowflakeStorageClient {
    *                required to decrypt/encrypt content in stage
    */
   public static SnowflakeGCSClient createSnowflakeGCSClient(
-      StageInfo stage, RemoteStoreFileEncryptionMaterial encMat, SFSession session)
+      StageInfo stage, RemoteStoreFileEncryptionMaterial encMat, SFSessionImpl session)
       throws SnowflakeSQLException {
     SnowflakeGCSClient sfGcsClient = new SnowflakeGCSClient();
     sfGcsClient.setupGCSClient(stage, encMat, session);
@@ -191,7 +189,7 @@ public class SnowflakeGCSClient implements SnowflakeStorageClient {
    */
   @Override
   public void download(
-      SFSession session,
+      SFSessionImpl session,
       String command,
       String localLocation,
       String destFileName,
@@ -356,7 +354,7 @@ public class SnowflakeGCSClient implements SnowflakeStorageClient {
    */
   @Override
   public InputStream downloadToStream(
-      SFSession session,
+      SFSessionImpl session,
       String command,
       int parallelism,
       String remoteStorageLocation,
@@ -577,7 +575,7 @@ public class SnowflakeGCSClient implements SnowflakeStorageClient {
    */
   @Override
   public void upload(
-      SFSession session,
+      SFSessionImpl session,
       String command,
       int parallelism,
       boolean uploadFromStream,
@@ -864,13 +862,13 @@ public class SnowflakeGCSClient implements SnowflakeStorageClient {
 
   @Override
   public void handleStorageException(
-      Exception ex, int retryCount, String operation, SFSession session, String command)
+      Exception ex, int retryCount, String operation, SFSessionImpl session, String command)
       throws SnowflakeSQLException {
     // no need to retry if it is invalid key exception
     if (ex.getCause() instanceof InvalidKeyException) {
       // Most likely cause is that the unlimited strength policy files are not installed
       // Log the error and throw a message that explains the cause
-      SnowflakeFileTransferAgent.throwJCEMissingError(operation, ex);
+      SnowflakeFileTransferAgentImpl.throwJCEMissingError(operation, ex);
     }
 
     if (ex instanceof StorageException) {
@@ -916,7 +914,7 @@ public class SnowflakeGCSClient implements SnowflakeStorageClient {
         if (se.getCode() == 401 && session != null && command != null) {
           // A 401 indicates that the access token has expired,
           // we need to refresh the GCS client with the new token
-          SnowflakeFileTransferAgent.renewExpiredToken(session, command, this);
+          SnowflakeFileTransferAgentImpl.renewExpiredToken(session, command, this);
         }
       }
     } else if (ex instanceof InterruptedException
@@ -1034,7 +1032,7 @@ public class SnowflakeGCSClient implements SnowflakeStorageClient {
    * @throws IllegalArgumentException when invalid credentials are used
    */
   private void setupGCSClient(
-      StageInfo stage, RemoteStoreFileEncryptionMaterial encMat, SFSession session)
+      StageInfo stage, RemoteStoreFileEncryptionMaterial encMat, SFSessionImpl session)
       throws IllegalArgumentException, SnowflakeSQLException {
     // Save the client creation parameters so that we can reuse them,
     // to reset the GCS client.
