+++ src/main/java/net/snowflake/client/jdbc/telemetry/TelemetryClient.java
@@ -15,6 +15,7 @@ import java.util.concurrent.Future;
 import net.snowflake.client.core.HttpUtil;
 import net.snowflake.client.core.ObjectMapperFactory;
 import net.snowflake.client.core.SFSession;
+import net.snowflake.client.core.SFSessionImpl;
 import net.snowflake.client.jdbc.SnowflakeConnectionV1;
 import net.snowflake.client.jdbc.SnowflakeSQLException;
 import net.snowflake.client.jdbc.telemetryOOB.TelemetryThreadPool;
@@ -42,7 +43,7 @@ public class TelemetryClient implements Telemetry {
   private final String serverUrl;
   private final String telemetryUrl;
 
-  private final SFSession session;
+  private final SFSessionImpl session;
   private LinkedList<TelemetryData> logBatch;
   private static final ObjectMapper mapper = ObjectMapperFactory.getObjectMapper();
 
@@ -53,7 +54,7 @@ public class TelemetryClient implements Telemetry {
   // false if meet any error when sending metrics
   private boolean isTelemetryServiceAvailable = true;
 
-  private TelemetryClient(SFSession session, int flushSize) {
+  private TelemetryClient(SFSessionImpl session, int flushSize) {
     this.session = session;
     this.serverUrl = session.getUrl();
 
@@ -92,7 +93,7 @@ public class TelemetryClient implements Telemetry {
    */
   public static Telemetry createTelemetry(Connection conn, int flushSize) {
     try {
-      return createTelemetry(conn.unwrap(SnowflakeConnectionV1.class).getSfSession(), flushSize);
+      return createTelemetry((SFSessionImpl) conn.unwrap(SnowflakeConnectionV1.class).getSfSession(), flushSize);
     } catch (SQLException ex) {
       logger.debug("input connection is not a SnowflakeConnection");
       return null;
@@ -115,7 +116,7 @@ public class TelemetryClient implements Telemetry {
    * @param session session to use for telemetry dumps
    * @return a telemetry connector
    */
-  public static Telemetry createTelemetry(SFSession session) {
+  public static Telemetry createTelemetry(SFSessionImpl session) {
     return createTelemetry(session, DEFAULT_FORCE_FLUSH_SIZE);
   }
 
@@ -126,7 +127,7 @@ public class TelemetryClient implements Telemetry {
    * @param flushSize maximum size of telemetry batch before flush
    * @return a telemetry connector
    */
-  public static Telemetry createTelemetry(SFSession session, int flushSize) {
+  public static Telemetry createTelemetry(SFSessionImpl session, int flushSize) {
     return new TelemetryClient(session, flushSize);
   }
